add -rename flag, for having different type in sherpadoc output than in source code
by default, types in the sherpadoc output are just the name as they occur in
the package. projects with multiple packages that hold the same type would
cause conflicting type names. by renaming one of them in the output, the source
can use the sensical names, and the sherpadoc can have nonconflicting names. a
rename could just prepend a package name to the final type name.